requirements.txt
customtkinter
Pillow
reportlab
pandas
openpyxl
python-barcode
matplotlib
arabic-reshaper
python-bidi
pyinstaller
